refactor: ToGoCode as method
