+++ openapigen/gocode.tmpl
